Create HTML/CSS and JS files
Built basic library display with a five-shelf bookcase, library sign,
window, and table with HTML and CSS to begin visualizing the library and
the books. Creatd a myLibrary array to contain book objects. Created the
book function using an object constructor to store basic book
information. Began coding for the addBookToLibrary function to add book
objects to the myLibrary array.