Fix form validation and add book to user interface
Created input:invalid CSS class to highlight empty inputs for user.
Updated the JavaScript code to show appropriate message for user when
filing out the form and to create a new "book" object from input values
and to create a visual representation of the book on the user interface
inside of the bookcase. Toggled the "bookcase" div element's visibility
using event listeners on the book catalog submit form.