Create Open Book modal and fix responsive layout
Made a new modal styled like an open book that appears when the user
clicks on the book object on the book case. Updated the color theme for
the body element's background to better match the "room". Updated the
column widths of child div elements within the field-content div
element to better fit within the modal content's boundaries.