Clear input fields and prevent array duplicates
Used find() array method to find any existing book objects in the array
that are an exact match to the input values and used if statements to
prevent adding the duplicate book to the myLibrary array and to inform
the user of the incident. Also reorganized the JavaScript code; created
the clearNotices and clearInput functions to allow different event
listeners to perform the same functions as necessary.