Include read toggle functionality on catalog form
Changed the method used to switch between "read" and "not read" by using
a function to toggle on the Book prototype instance. Applied a checkbox
input to the library catalog form and styled it to look like a toggle
switch. Updated the javascript code for the books' UI and used the
readInfo prototype method.